签到领金币  优化

+++ 51fbapi-biz/src/main/java/com/ald/fanbei/api/biz/service/AfTaskUserService.java
@@ -138,7 +138,7 @@ public interface AfTaskUserService {
      * @param couponIdList
      * @return
      */
-    BigDecimal getCouponAmountByIds(List<Long> couponIdList);
+    int getCouponAmountByIds(List<Long> couponIdList);
 
     /**
      * 批量更新用户状态
